Switch release step back to standard agents
This is a partial revert of commit 90deee4f84d2bf41d8b46ac439ad934605f557f3.

In that commit we changed the release step to run on an experimental
agent queue called "on-demand". That queue doesn't exist any more, so
releases are broken.

I've reverted it, but then made some additional changes. Back in 2021
the release step ran on our deploy agents, which had privileged access.
That's no longer a patter new use, so we want to:

* run the step on an standard unprivileged agent
* assume an IAM role via OIDC that has just the permissions we require
* use docker to define the build tools we need